Move project to library project, with a binary frontend. Also rename the project to sbl.
